refactor(openai-agents): split handler.ts into handler modules
handler.ts was 785 lines. Split into a handler/ folder parallel to the
codex layout:
  - state   — in-flight abort registry (getActiveAbort)
  - events  — RunItemStreamEvent → shared stream-state translation
              (handleRunItem + tool/message handlers + describeToolCall)
  - message — handleMessage entry point + flow-violation retry + the
              readRunUsage live-usage sampler

index.ts re-exports handleMessage / getActiveAbort. Importers (index,
factory, live-discovery test) updated to the barrel path.
